apns: refactor new api out of new
