+++ pypush/apns/lifecycle.py
@@ -13,19 +13,21 @@ from cryptography import x509
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import padding, rsa
 
-from . import _util, protocol, transport
+from . import protocol, transport, _util
+
 
 @asynccontextmanager
 async def create_apns_connection(
     certificate: x509.Certificate,
     private_key: rsa.RSAPrivateKey,
     token: typing.Optional[bytes] = None,
+    courier: typing.Optional[str] = None,
 ):
     async with anyio.create_task_group() as tg:
-        conn = Connection(tg, certificate, private_key, token)
+        conn = Connection(tg, certificate, private_key, token, courier)
         yield conn
-        tg.cancel_scope.cancel() # Cancel the task group when the context manager exits
-    await conn.aclose() # Make sure to close the connection after the task group is cancelled
+        tg.cancel_scope.cancel()  # Cancel the task group when the context manager exits
+    await conn.aclose()  # Make sure to close the connection after the task group is cancelled
 
 
 class Connection:
@@ -35,6 +37,7 @@ class Connection:
         certificate: x509.Certificate,
         private_key: rsa.RSAPrivateKey,
         token: typing.Optional[bytes] = None,
+        courier: typing.Optional[str] = None,
     ):
 
         self.certificate = certificate
@@ -45,6 +48,11 @@ class Connection:
         self._tg = task_group
         self._broadcast = _util.BroadcastStream[protocol.Command]()
         self._reconnect_lock = anyio.Lock()
+
+        if courier is None:
+            # Pick a random courier server from 1 to 50
+            courier = f"{random.randint(1, 50)}-courier.push.apple.com"
+        self.courier = courier
 
         self._tg.start_soon(self.reconnect)
         self._tg.start_soon(self._ping_task)
@@ -70,7 +78,7 @@ class Connection:
                 logging.warning("Closing existing connection")
                 await self._conn.aclose()
             self._conn = protocol.CommandStream(
-                await transport.create_courier_connection(courier="localhost")
+                await transport.create_courier_connection(courier=self.courier)
             )
             cert = self.certificate.public_bytes(serialization.Encoding.DER)
             nonce = (
@@ -100,14 +108,16 @@ class Connection:
             else:
                 assert ack.token == self.base_token
 
-
     async def aclose(self):
         if self._conn is not None:
             await self._conn.aclose()
         # Note: Will be reopened if task group is still running and ping task is still running
 
     T = typing.TypeVar("T", bound=protocol.Command)
-    async def receive_stream(self, filter: typing.Type[T], max: int = -1) -> typing.AsyncIterator[T]:
+
+    async def receive_stream(
+        self, filter: typing.Type[T], max: int = -1
+    ) -> typing.AsyncIterator[T]:
         async with self._broadcast.open_stream() as stream:
             async for command in stream:
                 if isinstance(command, filter):
@@ -115,7 +125,7 @@ class Connection:
                     max -= 1
                 if max == 0:
                     break
-            
+
     async def receive(self, filter: typing.Type[T]) -> T:
         async for command in self.receive_stream(filter, 1):
             return command
